project/fix/requirements.txt
ultralytics>=8.0.0
flask>=3.0.0
flask-cors>=4.0.0
Pillow>=10.0.0
roboflow>=1.1.0
torch>=2.0.0
